Make changes requested by the reviewer
